CocoaSpice: improved drawing latency
There is a BUG in OSX/iOS where TCP_NOPUSH is broken and causes a huge amount of latency. We patch spice server to not use it anymore and this makes everything nice again. See: http://lists.danga.com/pipermail/memcached/2005-March/001240.html

Additionally, the drawing is handled in the background and no longer as part of the invalidate event callback. This shouldn't be a real issue but it's good practice I guess.