keyboard: use hw keyboard when sw keyboard hidden
Refactored VMKeyboardView to VMKeyboardView and VMSoftKeyboardView. The
first is an invisible keyboard that works with hardware keyboard input.
This means you no longer have to show the keyboard to enter text. Hiding
the software keyboard will always toggle the hidden hardware keyboard.

Also disabled "pushing" the whole display upwards when the software keyboard
opens up. This wasn't really working with the two-keyboard-views and it
wasn't that good of an idea anyways. In the future, we should use private
APIs to make the software keyboard semi-transparent.