config: add local share directory selection
Remove global share directory selection (for now). Due to failure of testing
shared directories out locations outside the app sandbox does not actually
work. It requires a lot more work, which will be done in the future.

For now, we will remove that feature along with the built-in directory
picker and replace it with an in-app directory picker that only has access
inside the sandbox.

Resolves #263
Resolves #265