manager: removed force virtio-blk for virt machine
This is a BREAKING change to existing 'virt' machine types. Previously for
any devices that is marked as 'pflash' or 'sd' or any other type, it will
get forced to 'virtio' type if you select ARM/ARM64 and 'virt' machine.
Now, you must choose the 'virtio' type for the device interface.

Fixes #324