macOS: rework XPC helper with NSTask
Turns out you can only have one instance of a helper. Therefore we use NSTask
to launch multiple instances of QEMU. As a result, for macOS, we now embed the
QEMU binaries with the helper and launch them instead of using dylib. With this
change we also add a suffix to all the built dylibs in order to avoid collision
with macOS system libraries.