project: rename com.osy86.* to com.utmapp.*
This better reflects UTM's status as an open source project with multiple
contributers.

We also change from the x.ybz version format to x.y.z which Apple validation
requires. It also means that we put less emphasis on "beta" vs regular
builds. In reality, it is hard to determine if a build is stable since we do
not have dedicated QA testers.