CocoaSpice: crash when shared directory is invalid
spice-gtk uses g_get_user_special_dir(G_USER_DIRECTORY_PUBLIC_SHARE) as the
default path when shared-dir is not set. This evaluates to NULL on iOS and
later on the NULL pointer causes a crash when the WebDAV server is accessed.

We manually set the default share to /tmp/Public and put in a readme so the
user knows the files are temporary.

Fixes #528