config: remove orphaned drives on save or cancel
Currently, if the app crashes/quits in middle of a configuration, then there
could be orphaned drive images that has no config entry. This takes up space
and also prevents the user from importing a drive with the same name. We fix
this by adding deleting orphaned disk images for /tmp on save and for the VM
being configured on cancel. If it was an app crash, we do not automatically
delete but give the user a chance to adopt the drive image on the next config.
If they leave it unassigned, then it will be deleted on cancel.

Additionally, this allows users to unassign a drive image without deleting it
by setting the interface to 'none'.

Resolves #2275