CocoaSpice: removed extra code for connections
The main context is from a singleton so we do not need to pass
it from an higher level object. The sesssion created/ended
delegate methods are unneeded because the session can always be
created on init of CSConnection. Finally, there is no need to
start sharing in UTMSpiceIO as UTMVirtualMachines will always
start sharing if needed.