build: enable macOS build with provisioning profile
To support USB capture, we need to use a provisioning profile. To allow builds
to continue to work without a signing certificate, we leave out the signing
step in build_utm.sh and manually ad-hoc sign with the entitlements during the
package_mac.sh phase. We also strip out entitlements for local signed builds.