macOS: capture Cmd+Q and Cmd+W
If mouse is captured, we always pass Cmd+Any to the Metal view.
If mouse is not captured, we always warn for Cmd+Q.
If mouse is not captured, Cmd+W is blocked if VM is busy, or a confirm will
show if VM is running, or window will close if VM is stopped.

Resolves #2321