manager: rework QEMU -> QMP -> SPICE connect order
Previously, we would try to connect QMP and SPICE independently and wait
for a display to be created. This introduced a lot of race conditions
and is unworkable with GL scanout event being sent before we connect the
event. Now, we wait for QMP to be connected first (we still have to
retry because there's no good way to see if QEMU has started QMP yet).
Once QMP is connected, we start up SPICE. Once SPICE is connected, we
resume booting up the VM.