Change pending VM download progress from Progress to Float
This change is necessary because the previous way of doing it with Progress lead to a crash.
Previously, we used the URLSessionDownloadTask's progress property to get the Progress object and passed that to the view.
However this caused a crash within SwiftUI and Combine, likely due to the DownloadTask updating the Progress from a background queue.
This means we can't use a ProgressView in the UI. So I decided to simplify the ViewModel of UTMPendingVMView to just use a Float for the progress.