Fix downloadVM URL action
A file path confusion existed where the downloaded ZIP was extracted to the UTM
Documents directory instead of the temporary folder. As a result, UTM failed to import the downloaded file because a folder with the same name already existed. This is fixed by manually specifying the path for the unzip call.