Re-write UTM ZIP extraction with new library ZipFoundation
In the code review for the ZIP handling it was recommended to switch from unzipping everything that was downloaded to only selectively unzipping the parts that are actually the VM files. To implement this I had to change the ZIP library we use, as the old one did not support selective unzip of files.
This commit additionally introduces a duplicate VM filename check, so if you download a VM with the same name as one you already have, it will no longer silently ignore the new one, but save it with a (1) appended (or 2, 3, x depending on what's available).
It should be discussed that this duplicate name check either might exist as a system API or should be moved to UTMData, so it may be reused.