Fix #3388 invalid drive id for manually placed files
QEMU drive ids must start with a letter.
However if an image files was manually placed in the VM images folder,
this code path is triggered and a UUID set for the id.
This causes ids starting with numbers to be set sometimes.

This commit changes the name for these drives to use driveUUID naming to definitively work around name collisions.