downloader: re-written using Swift 5.5 concurrency
This addresses the race condition when starting a download with a tiny
file that completes quickly. The code is also simplified and the IPSW
and VM downloader are more unified.

The implementation wraps URLSessionDownloadDelegate methods into Task
and CheckedContinuation in order to adopt async/await semantics.
@MainActor is used for functions that access UI state which removes the
need for nested DispatchQueue.main.async calls everywhere.