manager: support running from shortcut
UTMData will now store bookmarks to VMs instead of path names. This is
for both sandboxed and external VMs. If a bookmark to an external VM is
created, `-accessShortcutWithCompletion:` will be called to create a
bookmark in QEMUHelper process (so it can access the data as well). The
QEMUHelper bookmark is stored in UTMViewState (along with shared
directory bookmark and removable drive bookmarks).

Because Virtualization.framework does not support cross-process
bookmarks, this means we cannot save shortcuts to Apple VMs. We also do
not support re-naming shortcut VMs because we do not have permission to
write to the parent directory (it is possible to manually rename the
file in Finder).

Resolves #132