display(iOS): remove toolbar/keyboard from viewState
With the new toolbar that is always visible, there is no value in
storing the legacy toolbar visibility in the per-VM viewState setting.
Now, legacy toolbar VMs will always start with the toolbar visible. With
that removed, we also remove the per-VM keyboard visibility setting to
be a global value (the keyboard visibility for the last VM launched will
persist). This also simplifies the code for the delegate and the VM
manager.