vm(qemu): serialize VM connection routine
Now that all VM operations are executed on a dedicated serialized queue,
it makes sense to have blocking operations on that queue when a VM is
starting up. This way if the caller makes multiple VM operations while
one is ongoing (start -> stop -> start), the next operation will block
on the queue. This also makes the logic easier to follow in vmStart.