qemu: support disabling PS/2 devices
This allows us to re-work how "legacy input" is handled. If "legacy
input" is enabled, then no USB bus will be created and default input
devices will be created (PS/2 on PC for example). If "legacy input" is
disabled, we now disable the PS/2 device in QEMU and only use USB mouse
and keyboard (previously this was an exception on devices like PC where
the PS/2 controller was always enabled). Note that we always use the USB
2.0 bus for input devices even if USB 3.0 is enabled.

Resolves #3674