config: new tweak "ForcePS2Controller"
On older versions, PS/2 controller is always enabled even when a USB
keyboard and mouse is used. This led to errors in some applications like
Windows 7 installer which may chose the wrong input device to use as the
primary one.

However, upon disabling the PS/2 controller by default, we found that a
"pc" target running Windows XP will now refuse to boot. In order to
address this we add a new tweak ForcePS2Controller which when set will
go back to the old behaviour of creating both a USB input device and
PS/2 input device. This will also be enabled by default for existing
VMs. New VMs created will have this tweak disabled by default.

Fixes #3723