display(macOS): synchronize caps lock
Make sure SPICE is aware of the caps lock state whenever flags modified
event is received (even if UTM is in the background). Also send the caps
lock state whenever the mouse is captured.

Note that if the guest OS is not booted up, the caps lock key event may be
dropped.

Fixes #3415