Support drive resizing
Add UTMLogToMemory to capture output of "qemu-img info", which we can parse the JSON output of in order to determine the virtual size of the drive.

Use the same widget style for resizing as we do for creating a drive (GB/MB), although require clicking "Resize" to do the resize itself.

Currently doesn't warn about having not clicked "resize" though.