UTM: use new configuration backend
The old backend was very limiting because of how it wrapped a dictionary into
an Objective-C class. It made working with SwiftUI awkward because of the
nested objects and lack of clean array support. It also makes using Swift
Concurrency difficult.

The new config backend unifies both the Apple and QEMU config formats which
were previously separate. This also allowed for much duplicate code to be
removed (particulary in the FS handling of disk images and load/saving).

A temporary UTMConfigurationWrapper class is implemented to interop with the
Objective-C backend mostly for QEMU support.