config: move network device to top of generated arguments
In some Linux distros, a "predictable" interface name means it uses the PCI
device index which can change when the user adds/removes displays, drives, etc.
By moving the network config to the front, we can ensure the PCI device index
of the network device is always first.

The downside is that existing VMs would need to reconfigure networking.