api: implement foundation for API client/server
Use UNIX domain sockets as the underlying communication interface between UTM
and the API clients. The application layer is JSON decoded into Swift structs.

A pure Swift socket implementation with support for Swift Concurrency powers
the client/server interface. A server is spawned as a detached Task and all
clients are handled in a TaskGroup.

The CLI interface is the first client for the API.