Added Hyper-V Enlightenments for x86_64 and i386
I added some Hyper-V Enlightenments [1] to the UI. Non of them are active by
default.

Enlightenments that require 'hv-vpindex' are omitted, since this enlightenment
is currently not available on macOS.

Also, Enlightenments that are used for nesting e.g. 'hv-evmcs', or only change
the Hyper-V Version ID 'hv_version-id-*' are omitted, as well.

Enlightenment 'hv-spinlocks' requires a parameter, currently 0x1fff is hardcoded
since this seems to be a sane value. [2]

This commit partially addresses #4914 but also suffers from issue #4940.

Currently this only works for newly created VMs.

[1] https://www.qemu.org/docs/master/system/i386/hyperv.html
[2] https://leduccc.medium.com/improving-the-performance-of-a-windows-10-guest-on-qemu-a5b3f54d9cf5#c5b9