config(apple): make trackpad usage a configurable option
Previously, we implemented both the USB pointer and trackpad device but it
introduced double-clicking and scrolling issues (#4636, #4645). After
setting the default to trackpad for macOS, it turns out that Monterey guests
do not have the right driver for it and does not work.

Now, we will attempt to detect Monterey on new installs and enable the
trackpad when it is not Monterey (no earlier macOS versions are supported).
Existing VMs will default to trackpad off for compatibility and will require
manual setting to enable it.

Fixes #5237