process: change current directory to socket destination
When we use AF_UNIX, the path has a maximum length of 104 characters. In the
past, we dependend on buggy behaviour that truncates the path length and both
QEMU and SPICE used the truncated path as the destination. This can result in
collisions between different VMs as well as other issues.

Now, we change the current working directory to where the AF_UNIX socket file
will reside and use relative addresses for both processes.