config(qemu): disable file locking on BIOS
In QEMU, when file locking is enabled and supported, the image
file will be opened with O_RDWR even if readonly=on. Due to some
race condition during QEMU startup, sometimes it detects that
locks are supported and in those times, the BIOS read from a read
only bookmark will fail.

Fixes #5523