display: implement multiple windows for visionOS
The current multi-window support for iOS uses UIWindowSceneDelegate and
requestSceneSessionActivation() to activate. For visionOS, we can use the
newer SwiftUI APIs with WindowGroups and openWindow(). While this can also
work on newer versions of iOS, for consistency we only use the new APIs on
visionOS.

Using the new APIs allow better handling of window states in visionOS.
Specifically, when we dynamically resize the VM window, the size does not
get saved and restored as the main window size on next launch.