qemu: disable locking on read-only images
When an ISO is mounted by macOS or if it is stored on an network drive that
does not support locking, "Failed to lock byte 100: Operation not supported"
shows up when you try to mount it. We disable locking in such cases and
implement a new QAPI interface for disabling locking with newly mounted
media as well.

Fixes #6564