config(qemu): default x86_64 CPU to highest supported by host
We cannot use "-cpu max" because it causes BSOD issues (see #2368). QEMU
documentation suggests choosing a CPU model that matches the generation of
the host. Luckily for us, XNU already did the difficult work of mapping the
CPUID to the generation. Note that Rosetta returns Westmere but AMD based
Hackintoshes will still be unsupported and will default to the old CPU
model.

This change is needed to both improve performance of virtualized CPUs on
Intel Macs as well as enable Windows 11 24H2 to be booted as it now requires
SSE4.2 support.

Fixes #6325
Fixes #6772