vm(qemu): load QEMU resources from Caches
For reasons still unknown, in the App Store build, we can no longer read the
BIOS from /Applications/UTM.app/Contents/Resources/qemu as a bookmark even
though this had worked fine in the past. As a workaround, we will now copy
all the data files to the App Sandbox's Cache and then share a bookmark to
that.

Resolves #6861