qemu: expose virtio-ramfb framebuffer in BAR
Windows requires a linear framebuffer to boot. We introduced virtio-ramfb
for this reason. Previously, when ramfb is activated, EDK2 will allocate
the framebuffer from system memory and pass the physical address to QEMU.
While this worked to boot Windows, upon installing the virtio-gpu
drivers, Windows does not unload the BDD driver for the POST display.
This led to various issues with a black display when Windows mistakingly
selects the ramfb display to target even after virtio-gpu is activated.

The cause of this issue is in dxgkrnl.sys. DpiFdoDetectPostDevice() is
called to check if the POST device can be retired after adding a new
minidisplay device. To determine if the POST display is owned by the
incoming device, its resource list is queried to see if the framebuffer
is located in any of the resource entries for this device. For a PCI
device, the resource list comes from the BARs.

Fixes #6332
Fixes #6883