When using Apple Virtualization notify guest OS of wake up events
Open a Virtio Socket for the guest OS to receive a wakeup command if
it connects. This could be the basis for more of a command structure,
like the QEMU guest agent. We are only using this to send a wakeup
command, however.

This is predominantly useful for the guest OS to resync time against
the hardware clock. Otherwise, using network time syncs relies on a
timer, which can take betweem 5 - 30 minutes before its next scheduled
sync. Also, network time syncs can sometimes fail with too much drift
to resync the local system clock. Yet, while ntp has a skew of 1000s
before failing, modern systemd Linux distros have switched to
system-timesyncd.service, so the author is unsure what max drift is
with this package.